adding intl for Predictions modal
