+++ src/components/index.js
@@ -1,3 +1,5 @@
 export { default as ZoeLogo } from './ZoeLogo';
 export { default as DevConsole, useDevConsole, DevConsoleProvider } from './DevConsole';
+export { default as OptionalFormattedMessage } from './OptionalFormattedMessage';
+
 export * from './forms';
